共通のコードをリファクタリング

+++ src/alphazero_learn.cpp
@@ -123,7 +123,7 @@ void alphaZero() {
 
         //1回目はmakeBatch内で十分棋譜が貯まるまで待ち時間が発生する.その生成速度を計算
         if (step_num == 1) {
-            double gen_speed = first_wait / (elapsedHours(start_time) * 3600);
+            float gen_speed = first_wait / (elapsedHours(start_time) * 3600);
             if (sleep_msec == -1) {
                 sleep_msec = (int64_t)(batch_size * 1000 / (batch_size_per_gen * gen_speed));
             }
